requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-groq
langchain-community
langchain-core
sentence-transformers
faiss-cpu
python-dotenv
pydantic
numpy
torch
transformers
requests
rich
tavily-python
cohere
langchain-huggingface
edge-tts
playwright